Added ability to sync individual gtasks immediately on save. Properties synced--all relevant task data e.g. title, due date, parent indentation. Properties ignored--ordering in a list. Also fixed several major bugs related to normal syncing.

+++ tests/src/com/todoroo/astrid/gtasks/GtasksIndentActionTest.java
@@ -139,7 +139,7 @@ public class GtasksIndentActionTest extends DatabaseTestCase {
         gtasksListService.updateLists(lists);
     }
 
-    private Task taskWithMetadata(int order, int indentation) {
+    private Task taskWithMetadata(long order, int indentation) {
         Task task = new Task();
         PluginServices.getTaskService().save(task);
         Metadata metadata = GtasksMetadata.createEmptyMetadata(task.getId());
