+++ tests/src/com/todoroo/astrid/gtasks/GtasksTaskMovingTest.java
@@ -223,10 +223,10 @@ public class GtasksTaskMovingTest extends DatabaseTestCase {
         gtasksTaskListUpdater.debugPrint("1");
     }
 
-    private void thenExpectMetadataOrderAndIndent(Task task, int order, int indent) {
+    private void thenExpectMetadataOrderAndIndent(Task task, long order, int indent) {
         Metadata metadata = gtasksMetadataService.getTaskMetadata(task.getId());
         assertNotNull("metadata was found", metadata);
-        assertEquals("order", order, (int)metadata.getValue(GtasksMetadata.ORDER));
+        assertEquals("order", order, metadata.getValue(GtasksMetadata.ORDER).longValue());
         assertEquals("indentation", indent, (int)metadata.getValue(GtasksMetadata.INDENT));
     }
 
@@ -265,7 +265,7 @@ public class GtasksTaskMovingTest extends DatabaseTestCase {
         return tasks;
     }
 
-    private Task createTask(String title, int order, int indent) {
+    private Task createTask(String title, long order, int indent) {
         Task task = new Task();
         task.setValue(Task.TITLE, title);
         PluginServices.getTaskService().save(task);
